Move parser-related files to own directory
